pyproject.toml
[project]
name = "m4lrg2025"
version = "0.1.0"
description = "Predicting multiple RNA profiles at once on Saccharomyces cerevisiae"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "biopython>=1.85",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.3",
    "torch>=2.7.1",
    "tqdm>=4.67.1",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
]

[tool.ruff]
src = ["src"]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = []
ignore = ["E501"]
